+++ src/Reactant/Jobs/IncrementReactionAggregatesJob.php
@@ -21,7 +21,7 @@ use Illuminate\Bus\Queueable;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Foundation\Bus\Dispatchable;
 
-final class IncrementAggregatesJob implements
+final class IncrementReactionAggregatesJob implements
     ShouldQueue
 {
     use Dispatchable;